+++ ChangeLog.md
@@ -1,5 +1,8 @@
 ## change log ##
 
+* 2016-06-18: 0.3.2:
+  - extract clipboard-copy utility into a directive
+
 * 2016-06-17: 0.3.2:
   - \#30 "dispatch term request"
     - initial preparations to dispatch given URI (rUri) so it uses the existing 'ont' display,